+++ CHANGELOG.md
@@ -9,8 +9,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keep-a-changelog], and this project adheres to [Semantic Versioning v2.0.0][semver]. See [the README file][docs.readme] for more details on how it is versioned.
 
-## [Unreleased]
+## [1.10.0] - 2026-02-13
+### Changed
+- [#14](https://github.com/ForNeVeR/Generaptor/pull/14): the job generator now verifies that `runs-on` is specified.
+
 ### Added
+- [#117](https://github.com/ForNeVeR/Generaptor/issues/117): `verify` command will now output diff in patch format.
+- Support for `working-directory` at the `step` level.
 - Support for `timeout-minutes` at the job level via the `jobTimeout` command.
 
 ## [1.9.1] - 2026-01-24
@@ -99,5 +104,6 @@ The initial release of this package. Main features:
 [1.7.0]: https://github.com/ForNeVeR/Generaptor/compare/v1.6.1...v1.7.0
 [1.8.0]: https://github.com/ForNeVeR/Generaptor/compare/v1.7.0...v1.8.0
 [1.9.0]: https://github.com/ForNeVeR/Generaptor/compare/v1.8.0...v1.9.0
-[1.9.0]: https://github.com/ForNeVeR/Generaptor/compare/v1.9.0...v1.9.1
-[Unreleased]: https://github.com/ForNeVeR/Generaptor/compare/v1.9.1...HEAD
+[1.9.1]: https://github.com/ForNeVeR/Generaptor/compare/v1.9.0...v1.9.1
+[1.10.0]: https://github.com/ForNeVeR/Generaptor/compare/v1.9.1...v1.10.0
+[Unreleased]: https://github.com/ForNeVeR/Generaptor/compare/v1.10.0...HEAD